% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catGUI.R
\name{catGUI}
\alias{catGUI}
\title{Graphical User Interface}
\usage{
catGUI()
}
\description{
\code{catGUI} creates a shiny app for \code{cat.sim}
}
\details{
Imported item files and theta files must be comma-delimited and have headers
}

