% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module4_cat_helpers.R
\name{catHelpers}
\alias{catHelpers}
\alias{cat_select_randomesque}
\title{CAT Helper Functions}
\usage{
cat_select_randomesque(theta, pool, randomesque)
}
\arguments{
\item{theta}{the current theta estimate}

\item{pool}{the item pool}

\item{randomesque}{the randomesque parameter}
}
\description{
CAT Helper Functions

\code{cat_select_randomesque} selects an item from k most informative ones
}

