% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_ctt_utils.R
\name{cronbach_alpha}
\alias{cronbach_alpha}
\title{Cronbach's alpha}
\usage{
cronbach_alpha(responses)
}
\arguments{
\item{responses}{the oberved responses, 2d matrix}
}
\description{
\code{cronbach_alpha} computes Cronbach's alpha internal consistency reliability
}
\examples{
cronbach_alpha(model_3pl_gendata(1000, 20)$u)
}
