% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1_helpers.R
\name{estimate_helpers}
\alias{estimate_helpers}
\alias{estimate_nr_iteration}
\alias{model_polytomous_3dindex}
\alias{model_polytomous_3dresponse}
\title{Helper functions of Model Estimation}
\usage{
estimate_nr_iteration(param, free, dv, h_max, lr, bound)

model_polytomous_3dindex(u)

model_polytomous_3dresponse(u)
}
\arguments{
\item{param}{the parameter being estimated}

\item{free}{TRUE to freely estimate specific parameters}

\item{dv}{the first and second derivatives}

\item{h_max}{the maximum value of h}

\item{lr}{the learning rate}

\item{bound}{the lower and upper bounds of the parameter}

\item{u}{the observed response, 2d matrix, values start from 0}
}
\description{
miscellaneous helper functions for estimating IRT models

\code{estimate_nr_iteration} updates the parameters using the newton-raphson method
}
\keyword{internal}
