% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2_estimation.R
\name{estimate_people}
\alias{estimate_people}
\alias{estimate_people_3pl_eap}
\alias{estimate_people_3pl_map}
\alias{estimate_people_3pl_mle}
\title{Estimate People Parameters}
\usage{
estimate_people(responses, items, model = "3pl", method = "mle", ...)

estimate_people_3pl_mle(responses, items, init = 0, iter = 30,
  conv = 0.01, bound = 3.5, debug = FALSE)

estimate_people_3pl_map(responses, items, prior.mean = 0, prior.sd = 1,
  init = 0, iter = 15, conv = 0.01, bound = 3.5, debug = FALSE)

estimate_people_3pl_eap(responses, items)
}
\arguments{
\item{responses}{a data frame or matrix of responses}

\item{items}{a data frame of item parameters}

\item{model}{the IRT model}

\item{method}{the estimation method}

\item{...}{further arguments}

\item{init}{initial values of people parameters, a number or a vector}

\item{iter}{the maximum of Newton-Raphson iterations}

\item{conv}{convergence criterion}

\item{bound}{the bounds of estimated people parameters}

\item{debug}{\code{TRUE} to turn on debugging mode}

\item{prior.mean}{the mean of the prior distribuiton}

\item{prior.sd}{the standard deviation of the prior distribution}
}
\value{
a data frame of estimated people parameters
}
\description{
Estimate people parameters when item parameters are known

\code{estimate_people_3pl_mle} is the maximum likelihood (ML) estimator of the people parameter.

\code{estimate_people_3pl_map} is the maximum a posteriori (MAP) estimator of people parameters

\code{estimate_people_3pl_eap} is the expected a priori (EAP) estimator of people parameters
}
\details{
The debug mode draws a convergence curve. \cr
For the maximum likelihood (ML) estimation, refer to Baker and Kim (2004), pp. 66-69.

For the maximum a posteriori (MAP) estimation, refer to Baker and Kim (2004), pp. 192.

For the expected a priori (EAP), refer to Baker and Kim (2004), pp. 193.
}
\examples{
\dontrun{
# data generation
data <- irt_model("3pl")$gendata(500, 50)
# MLE
x <- estimate_people(data$responses, data$items, "3pl", "mle", debug=TRUE)
cor(data$people$theta, x$people$theta)
plot(data$people$theta, x$people$theta, xlim=c(-4,4), ylim=c(-4,4), 
xlab="True", ylab="Est.", col=rgb(.8,.2,.2,.5), pch=16)
# MAP
x <- estimate_people(data$responses, data$items, "3pl", "map", debug=TRUE)
cor(data$people$theta, x$people$theta)
plot(data$people$theta, x$people$theta, xlim=c(-4,4), ylim=c(-4,4), 
xlab="True", ylab="Est.", col=rgb(.8,.2,.2,.5), pch=16)
# EAP
x <- estimate_people(data$responses, data$items, "3pl", "eap")
cor(data$people$theta, x$people$theta)
plot(data$people$theta, x$people$theta, xlim=c(-4,4), ylim=c(-4,4), 
xlab="True", ylab="Est.", col=rgb(.8,.2,.2,.5), pch=16)
}
}

