% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2_estimation_helpers.R
\name{estimationHelper}
\alias{estimate_items_3pl_check_input}
\alias{estimate_items_3pl_init_par}
\alias{estimate_people_3pl_check_input}
\alias{estimationHelper}
\alias{quadrature}
\title{Estimation Helper Functions}
\usage{
estimate_people_3pl_check_input(responses, items)

quadrature()

estimate_items_3pl_check_input(responses, people)

estimate_items_3pl_init_par(fix, init, n.items)
}
\arguments{
\item{responses}{the responses data}

\item{items}{the item parameters}

\item{people}{the people parameters}

\item{fix}{a list of fixed parameter values}

\item{init}{the initial values of parameters}

\item{n.items}{the number of items}
}
\description{
Estimation Helper Functions
}
\details{
\code{qudrature} was originally intended to call \code{gaussquad::hermite.h.quadrature.rules(12)[[12]]}
}

