% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_irt_utils.R
\name{expected_raw_score_dist}
\alias{expected_raw_score_dist}
\title{#' Distribution of Expected Raw Scores}
\usage{
expected_raw_score_dist(t, a, b, c)
}
\arguments{
\item{t}{the ability parameters, 1d vector}

\item{a}{the item discrimination parameters, 1d vector}

\item{b}{the item difficulty parameters, 1d vector}

\item{c}{the item guessing parameters, 1d vector}
}
\description{
Calculate the distribution of expected raw scores
}
