% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1_stat_utils.R
\name{freq}
\alias{freq}
\title{Count Frequency}
\usage{
freq(x, val)
}
\arguments{
\item{x}{a vector being counted}

\item{val}{a vector of valid counting values}
}
\description{
Count (cumulative) frequency and percentage using user-defined values
}
\examples{
freq(sample(1:4, 150, replace=TRUE), 5:1)
freq(sample(1:4, 150, replace=TRUE), 1:3)
}

