% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_gui_estimation.R
\name{gui.estimation}
\alias{gui.estimation}
\title{Graphical User Interface for Parameter Estimation}
\usage{
gui.estimation()
}
\description{
\code{gui.estimation} is a shiny app for parameter estimation
}

