% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_stat_utils.R
\name{hermite_gauss}
\alias{hermite_gauss}
\title{Hermite-Gauss Quadrature}
\usage{
hermite_gauss(degree = c("20", "11", "7"))
}
\arguments{
\item{degree}{the degree of hermite-gauss quadrature: '20', '11', '7'}
}
\description{
Pre-computed hermite gaussian quadratures points and weights
}
\keyword{internal}
