% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_misc.R
\name{misc}
\alias{misc}
\alias{cronbach_alpha}
\alias{summed_score_dist}
\title{Miscellaneous}
\usage{
cronbach_alpha(u)

summed_score_dist(t, a, b, c)
}
\arguments{
\item{u}{the oberved responses, a matrix}

\item{t}{the ability parameters, a vector}

\item{a}{the item discrimination parameters, a vector}

\item{b}{the item difficulty parameters, a vector}

\item{c}{the item guessing parameters, a vector}
}
\description{
Miscellaneous

\code{cronbach_alpha} computes the Cronbach's alpha

\code{summed_score_dist} computes the distribution of summed scores
}
\examples{
cronbach_alpha(model_3pl_gendata(1000, 20)$u)
cronbach_alpha(model_3pl_gendata(1000, 40)$u)
}
