% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module5_mst_helpers.R
\name{mstHelpers}
\alias{mstHelpers}
\alias{mst_indices_input}
\alias{mst_reindex_routes}
\title{MST Helper Functions}
\usage{
mst_reindex_routes(x)

mst_indices_input(x, indices)
}
\arguments{
\item{x}{a \code{mst} object}

\item{indices}{the route or module indices}
}
\description{
MST Helper Functions

\code{mst_reindex_routes} sorts an reindexes the routes map

\code{mst_indices_input} converts indices input to indices of routes in topdown method or modules in bottomup method
}

