% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module2_estimations.R
\name{estimate.people.3pl.check.input}
\alias{estimate.items.3pl.check.input}
\alias{estimate.items.3pl.init.par}
\alias{estimate.people.3pl.check.input}
\alias{quadrature}
\title{Helper Functions for Parameter Estimation}
\usage{
estimate.people.3pl.check.input(responses, items)

quadrature()

estimate.items.3pl.check.input(responses, people)

estimate.items.3pl.init.par(fix, init, n.items)
}
\arguments{
\item{responses}{the responses data}

\item{items}{the item parameters}

\item{people}{the people parameters}

\item{fix}{the fixed values of parameters}

\item{init}{the initial values of parameters}

\item{n.items}{the number of items}
}
\value{
a list of validated \code{responses} and \code{items}

a list of validated \code{responses} and \code{people}

a list of three initial parameters and three boolean vectors (whether fixed or not)
}
\description{
\code{estimate.people.3pl.check.input} is a helper function for validating inputs

\code{quadrature} is a helper fucntion for getting hermite gauss quadrature points

\code{estimate.items.3pl.check.input} is a helper function for validating inputs

\code{estimate.items.3pl.init.par} is a helper function for initiating item parameters
}
\details{
\code{qudrature} was originally intended to call \code{gaussquad::hermite.h.quadrature.rules(12)[[12]]}
}

