% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_stat_utils.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{a numeric vector/matrix}

\item{y}{a numeric vector/matrix}
}
\description{
Root mean squared error (RMSE) of two numeric vectors/matrices
}
