% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_ctt_utils.R
\name{spearman_brown}
\alias{spearman_brown}
\alias{spearman_brown_reverse}
\title{Spearman Brown Prophecy}
\usage{
spearman_brown(n, rho)

spearman_brown_reverse(rho, target)
}
\arguments{
\item{n}{extend the test length to n-fold}

\item{rho}{the reliability of current test}

\item{target}{the targeted reliability}
}
\description{
Use Spearman-brown formula to compute the predicted reliability
when the test length is extened to n-fold or reversely the n-fold extension of
test length in order to reach the targeted reliability
}
\examples{
spearman_brown(2, .70)
spearman_brown_reverse(.70, .85)
}
