% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_utils.R
\name{utils}
\alias{utils}
\alias{rmse}
\alias{freq}
\title{Utility functions}
\usage{
rmse(x, y)

freq(x, values = NULL)
}
\arguments{
\item{x}{a vector/matrix of numeric values}

\item{y}{a vector/matrix of numeric values}

\item{values}{a vector of valid values, \code{NULL} for all values}
}
\description{
Utility functions

\code{rmse} computes the root mean squared error of two numeric vectors/matrices

\code{freq} computes the frequency and percentage of given values
}
\examples{
rmse(1 + rnorm(100), 2 + rnorm(100))
freq(sample(1:5, 100, replace=TRUE))
}
