% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module3_ata_helpers.R
\name{ataHelpers}
\alias{ataHelpers}
\alias{ata_append}
\alias{ata_form_index}
\alias{ata_obj_coef}
\title{ATA Helper Functions}
\usage{
ata_append(x, mat, dir, rhs)

ata_form_index(x, forms, collapse)

ata_obj_coef(x, coef, compensate)
}
\arguments{
\item{x}{the ata object}

\item{mat}{the coefficient}

\item{dir}{direction}

\item{rhs}{right-hand-side value}

\item{forms}{the forms indices}

\item{collapse}{\code{TRUE} to collapse forms}

\item{coef}{the coefficients}

\item{compensate}{\code{TRUE} to combine coefficients}
}
\description{
ATA Helper Functions

\code{ata_append} adds constraint data into ata

\code{ata_form_index} converts input forms into actual form indices in LP

\code{ata_obj_coef} processes input coefficients for setting objective functions
}

