% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module0_stat_utils.R
\name{freq}
\alias{freq}
\title{Frequency Counts}
\usage{
freq(x, values = NULL, rounding = NULL)
}
\arguments{
\item{x}{a numeric or character vector}

\item{values}{valid values, \code{NULL} to include all values}

\item{rounding}{round percentage to n-th decimal places}
}
\description{
Frequency counts of a vector
}
