% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module5_mst_simulation.R
\name{mst_sim}
\alias{mst_sim}
\title{MST Simulation}
\usage{
mst_sim(x, theta.true, rdp = NULL)
}
\arguments{
\item{x}{an assembled mst}

\item{theta.true}{the true value of theta parameter}

\item{rdp}{a list of routing decision points}
}
\description{
\code{mst_sim} runs a MST simulation
}

