% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1_utilities.R
\name{rmse}
\alias{freq}
\alias{rmse}
\title{Utility Functions}
\usage{
rmse(x, y)

freq(x, values)
}
\arguments{
\item{x}{a vector/matrix of numeric values}

\item{y}{a vector/matrix of numeric values}

\item{values}{a vector of valid counting values}
}
\description{
\code{rmse} computes the root mean squared error (rsme) of two numeric vectors/matrices

\code{freq} counts the frequency using user-defined values
}
\examples{
rmse(1:10 + rnorm(10), 1:10 + rnorm(10))
rmse(matrix(1:100 + rnorm(100), ncol=5), matrix(1:100 + rnorm(100), ncol=5))
freq(sample(1:4, 100, replace=TRUE), 5:1)
freq(sample(1:4, 100, replace=TRUE), 1:5)
}
\seealso{
Other res: \code{\link{cat.sim}}
}

