% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module1_stat_utils.R
\name{rmse}
\alias{rmse}
\title{Compute Root Mean Square Error}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{a vector/matrix of numeric values}

\item{y}{a vector/matrix of numeric values}
}
\description{
Compute the root mean squared error (rsme) of two numeric vectors/matrices
}

